完成了printf、scanf、tty、成功链接了整个系统

+++ os_lab3/test/libc/string.cpp
@@ -107,7 +107,7 @@ void reverse(char* buffer)
 {
     char ch;
     size_t len = strlen(buffer);
-    for(int i = 0; i < len / 2; ++i)
+    for(size_t i = 0; i < len / 2; ++i)
     {
         ch = *(buffer + i);
         *(buffer + i) = *(buffer + len - 1 - i);
